Get rid of the local fedora repo
This repo must be creted manually, which is not good for the
"one click" make qubes target.

The rpms for this repo are downloaded by revisor anyway from the
fedora repo and are cached in build/work/revisor-yumchace anyway,
so subsequent make iso attempts do not need to download them again.