comps: Do not install grubby
grubby provides the new-kernel-pkg script, which overrides the behaviour
of the kernel-install script. In addition, grubby causes the following
line

  xen_rm_opts=

to become

  xen_rm_opts

in the GRUB2 configuration file (/boot/grub2/grub.conf) every time
it is invoked, where the trailing equals sign ('=') is removed.
This causes a syntax error to be printed out at every boot.