lorax: Remove branding code
This code caused transient build failures when there is more than one
package that provides /etc/system-release _available_ in any of the
repositories, irrespective of _any_ dependecies whatsoever. The failure
is non-deterministic, because depends on the order of packages returned
by search query.

Automatic branding install is stupid anyway and should instead be done
by package groups, kickstarts and/or whatever. But it does not surprise
me at all, because this is package is maintained as part of Fedora/RH.

fixes QubesOS/qubes-issues#2076