anaconda: really set default paritioning layout
setDefaultPartitioning is called when executing kickstart "autopart"
command - which is the case if one choose to not change disk
partitioning. But in manual partitioning there is also an option to
start with automatic layout and in that case setDefaultPartitioning
isn't called at all. Which results in failed partitioning (missing '/'
and bootable partition).

Similar thing is already fixed in master commit
378cfc44dd218c61b838f4f9011dcfd790df59eb.

QubesOS/qubes-issues#2294
Fixes QubesOS/qubes-issues#3334