Set SOURCE_DATE_EPOCH for the build as last commit time
This isn't exactly "last modification time of source", as the source for
the final ISO include also all rpm packages from other components (and
from Fedora), which may be newer than installer-qubes-os itself. But in
practice it shouldn't be a problem as most of them will be only included
as files without relying on any specific timestamps inside.