qubes-release: provide system-release(releasever)
dnf do not support distroverpkg option to detect $relasever value.
Instead, it looks for a various packages, starting with
'system-release(releasever)', then 'system-release' and
'redhat-release'.
Until Fedora 29's dnf it took package version for $releasever value, but
starting with Fedora 29 it takes Provides value instead (it looks like
that code was in place before already, but for some reason it didn't worked).
Both 'redhat-release' and 'system-release' are set to base Fedora
version and needs to stay this way to satisfy Fedora's packages
dependencies. This means, $releasever was set to base Fedora version,
not Qubes version, which breaks repository urls.
Add 'system-release(releasever)' (which is checked first) with Qubes
version to solve this problem.

Fixes QubesOS/qubes-issues#4477